[refactor] 프로필쪽 코드 리팩토링
